# docs/openapi.yaml
openapi: 3.0.3
info:
  title: The Indian Startup API
  description: |
    Comprehensive API for managing user accounts, product access, progress tracking, and portfolio management.
    
    ## Authentication
    All authenticated endpoints require a valid session through NextAuth.js.
    
    ## Rate Limiting
    - Authenticated endpoints: 100 requests/minute per user
    - Public endpoints: 20 requests/minute per IP
    
    ## Support
    For API support, contact: support@theindianstartup.in
  version: 1.2.0
  contact:
    name: The Indian Startup Support
    email: support@theindianstartup.in
    url: https://theindianstartup.in
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://theindianstartup.in/api
    description: Production server
  - url: https://test.theindianstartup.in/api
    description: Test server
  - url: http://localhost:3000/api
    description: Development server
paths:
  /user/profile:
    get:
      summary: Get user profile
      description: Retrieve the current user's profile information
      tags:
        - User Management
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    put:
      summary: Update user profile
      description: Update the current user's profile information
      tags:
        - User Management
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session token
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        details:
          type: string
          description: "Additional error details (development only)"

    # User schemas
    UserProfile:
      type: object
      properties:
        id:
          type: string
          example: "user_123"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        totalXP:
          type: integer
          example: 1250
        currentStreak:
          type: integer
          example: 7
        hasCompletedOnboarding:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
    UserProfileResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserProfile'
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          example: "John Doe"
        bio:
          type: string
          example: "Startup founder"
        website:
          type: string
          format: uri
          example: "https://example.com"

    # Dashboard schemas
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Authentication required"
...